services/vix_monitor/requirements.txt
yfinance==0.2.37
boto3==1.34.0
psycopg2-binary==2.9.9
